+++ include/flang/lower/ConvertType.h
@@ -17,7 +17,7 @@
 ///
 /// [Coding style](https://llvm.org/docs/CodingStandards.html)
 
-#include "../../../lib/common/Fortran.h"
+#include "flang/common/Fortran.h"
 #include "mlir/IR/Types.h"
 
 namespace mlir {
