+++ lib/lower/ASTBuilder.cpp
@@ -7,7 +7,7 @@
 //===----------------------------------------------------------------------===//
 
 #include "flang/lower/ASTBuilder.h"
-#include "../parser/parse-tree-visitor.h"
+#include "flang/parser/parse-tree-visitor.h"
 #include <cassert>
 #include <utility>
