+++ lib/lower/ConvertExpr.cpp
@@ -7,20 +7,20 @@
 //===----------------------------------------------------------------------===//
 
 #include "flang/lower/ConvertExpr.h"
-#include "../common/default-kinds.h"
-#include "../common/unwrap.h"
-#include "../evaluate/fold.h"
-#include "../evaluate/real.h"
-#include "../semantics/expression.h"
-#include "../semantics/symbol.h"
-#include "../semantics/type.h"
 #include "fir/Dialect/FIRDialect.h"
 #include "fir/Dialect/FIROps.h"
 #include "fir/Dialect/FIRType.h"
+#include "flang/common/default-kinds.h"
+#include "flang/common/unwrap.h"
+#include "flang/evaluate/fold.h"
+#include "flang/evaluate/real.h"
 #include "flang/lower/Bridge.h"
 #include "flang/lower/ConvertType.h"
 #include "flang/lower/OpBuilder.h"
 #include "flang/lower/Runtime.h"
+#include "flang/semantics/expression.h"
+#include "flang/semantics/symbol.h"
+#include "flang/semantics/type.h"
 #include "mlir/Dialect/AffineOps/AffineOps.h"
 #include "mlir/Dialect/LLVMIR/LLVMDialect.h"
 #include "mlir/Dialect/StandardOps/Ops.h"
