+++ tools/bbc/bbc.cpp
@@ -11,27 +11,27 @@
 ///
 //===----------------------------------------------------------------------===//
 
-#include "../../lib/common/Fortran-features.h"
-#include "../../lib/common/default-kinds.h"
-#include "../../lib/parser/characters.h"
-#include "../../lib/parser/dump-parse-tree.h"
-#include "../../lib/parser/message.h"
-#include "../../lib/parser/parse-tree-visitor.h"
-#include "../../lib/parser/parse-tree.h"
-#include "../../lib/parser/parsing.h"
-#include "../../lib/parser/provenance.h"
-#include "../../lib/parser/unparse.h"
-#include "../../lib/semantics/expression.h"
-#include "../../lib/semantics/semantics.h"
-#include "../../lib/semantics/unparse-with-symbols.h"
 #include "fir/Dialect/FIRDialect.h"
+#include "fir/InternalNames.h"
+#include "fir/KindMapping.h"
+#include "flang/common/Fortran-features.h"
+#include "flang/common/default-kinds.h"
 #include "flang/lower/Bridge.h"
 #include "flang/lower/ConvertExpr.h"
 #include "flang/optimizer/CodeGen/CodeGen.h"
-#include "flang/optimizer/InternalNames.h"
-#include "flang/optimizer/KindMapping.h"
 #include "flang/optimizer/Transforms/Passes.h"
 #include "flang/optimizer/Transforms/StdConverter.h"
+#include "flang/parser/characters.h"
+#include "flang/parser/dump-parse-tree.h"
+#include "flang/parser/message.h"
+#include "flang/parser/parse-tree-visitor.h"
+#include "flang/parser/parse-tree.h"
+#include "flang/parser/parsing.h"
+#include "flang/parser/provenance.h"
+#include "flang/parser/unparse.h"
+#include "flang/semantics/expression.h"
+#include "flang/semantics/semantics.h"
+#include "flang/semantics/unparse-with-symbols.h"
 #include "mlir/Conversion/LoopToStandard/ConvertLoopToStandard.h"
 #include "mlir/IR/MLIRContext.h"
 #include "mlir/IR/Module.h"
